+++ src/com/pos/Data/Entities/Goods.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class Goods {
 
-    private int goodsType;
+    private long goodsType;
     private int quantityGoods;
     private String imageName;
     private String publicName;
@@ -41,7 +41,7 @@ public class Goods {
 
     private static Goods createGoodsFromShortString(String strGoodsQuantity) {
         String[] arrayGoodsQuantity = strGoodsQuantity.split("\\\\");
-        Goods goods = new Goods(Integer.parseInt(arrayGoodsQuantity[0]), Integer.parseInt(arrayGoodsQuantity[1]),
+        Goods goods = new Goods(Long.parseLong(arrayGoodsQuantity[0]), Integer.parseInt(arrayGoodsQuantity[1]),
                 null,null,null,null,null,null,null);
         return goods;
     }
